refactor: 지원서 관련 타입 types/application.ts 로 이동
