--- src/types/apis/uploadFile.ts
@@ -1,11 +1,0 @@
-export interface PresignedUrlRequest {
-  name: string;
-  contentType: string;
-  contentLength: number;
-}
-
-export interface PresignedUrlResponse {
-  cdnUrl: string;
-  presignedUrl: string;
-  expiration: string;
-}
